src/pyproject.toml
[tool.pdm]
distribution = false

[tool.pdm.build]
includes = []
[project]
name = "core-app"
version = "0.1.0"
description = ""
authors = [
    {name = "akuznetzz", email = "akuznetzz@yandex.ru"},
    {name = "Alex Kuznetsov"},
]
dependencies = [
    "django<6.0.0,>=5.0.1",
    "djangorestframework<4.0.0,>=3.14.0",
    "psycopg[binary]<4.0.0,>=3.1.17",
    "dynamic-rest @ file:///${PROJECT_ROOT}/dynamic-rest-master",
    "pre-commit>=3.6.0",
    "basedmypy>=2.4.0",
    "reorder-python-imports>=3.12.0",
    "ruff>=0.1.15",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.3.1",
]
requires-python = ">=3.12,<4.0"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.mypy] # https://mypy.readthedocs.io/en/stable/config_file.html
disable_error_code = ['explicit-override']
explicit_package_bases = true
implicit_reexport = true  # https://github.com/konradhalas/dacite/issues/133#issuecomment-1186549192
no_implicit_optional = true
ignore_missing_imports = true
follow_imports = "skip"
disallow_untyped_defs = true
disallow_any_expr = false
disallow_any_explicit = false
disallow_any_decorated = false

disallow_subclassing_any = false
disallow_any_unimported = false
disallow_untyped_calls = false

exclude = [
    '\.git',
    '\.cache',
    '\.mypy_cache',
    '\.pytest_cache',
    '\.prospector',
    '\.gitlab',
    'charts',
    'metabase',
    'docs',
    'grafana',
    '\.idea',
    '__pycache__\/',
    'migrations\/',
    "venv",
    "dynamic-rest-master"
]

[tool.ruff]
select = ["E", "F", "B", "Q", "T", "N", "UP", "W", "C90", "COM", "PYI", "RSE", "RET", "SIM", "TID", "TCH", "ARG", "PT",
    "PTH", "ERA", "PGH", "PL", "TRY", "RUF", "ANN", "S", "BLE", "C4", "PIE",
    # "D" – будет требовать docstrings для публичных классов, методов и функций
    # "INP" – будет сообщать, что модуль расположен в каталоге, где нет init.py. — Спорно...
    # "G" – запретит использовать форматирование в при логгинге
    # "DTZ" – утомит требованиями таймзоны при манипуляциях с дататми
]

line-length = 120
ignore = ["RUF001", "RUF003", "D203", "D213", "ANN101", "ANN102", "ANN401", "S101", "ANN201", "ANN003"]
exclude = ["migrations", "dynamic-rest-master", "tests", "manage.py"]

[tool.ruff_wrapper]
baseline_file = './baselines/ruff'

[tool.ruff.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"

[tool.ruff.per-file-ignores]
"tests/**/test_*.py" = [
     "ANN001", "PGH004", "ANN201", "PLR0913", "BLE001",
]
